requirements.txt
Flask==2.0.1
Flask-JWT-Extended==4.6.0
Flask-RESTx==0.5.1
Flask-Marshmallow==0.15.0
Flask-CORS==4.0.0
SQLAlchemy==1.4.49
python-dotenv==1.0.0
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.0.1
pytest==7.0.0
requests==2.31.0
beautifulsoup4==4.12.0
Werkzeug==2.0.1